requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudevents==1.11.0
deprecation==2.1.0
flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
packaging==25.0
playwright==1.52.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyee==13.0.0
requests==2.32.3
rsa==4.9.1
slack-sdk==3.35.0
typing-extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
werkzeug==3.1.3
